Radio button refactor
