+++ @utahdts/utah-design-system/index.js
@@ -55,8 +55,7 @@ export { default as Popup } from './react/components/popups/Popup';
 export { default as PopupMenuShape } from './react/propTypesShapes/header/PopupMenuShape';
 export { default as popupPlacement } from './react/enums/popupPlacement';
 export { default as rectContainsPoint } from './react/util/rectContainsPoint';
-export { RadioButtonsWrapper } from './react/components/forms/radioButton/RadioButtonsWrapper';
-export { RadioButton } from './react/components/forms/radioButton/RadioButton';
+export { RadioButton } from './react/components/forms/RadioButton';
 export { default as RefShape } from './react/propTypesShapes/RefShape';
 export { default as SegmentedButton } from './react/components/buttons/SegmentedButton';
 export { default as Select } from './react/components/forms/Select';
