+++ @utahdts/utah-design-system/react/components/forms/RadioButton.jsx
@@ -1,13 +1,16 @@
 // @ts-check
 import PropTypes from 'prop-types';
-import React, { useCallback } from 'react';
-import { useRadioButtonsContext } from './hooks/useRadioButtonsContext';
-import RefShape from '../../../propTypesShapes/RefShape';
-import RequiredStar from '../RequiredStar';
-import joinClassNames from '../../../util/joinClassNames';
+import React from 'react';
+import RefShape from '../../propTypesShapes/RefShape';
+import RequiredStar from './RequiredStar';
+import joinClassNames from '../../util/joinClassNames';
+import useCurrentValuesFromForm from '../../hooks/forms/useCurrentValuesFromForm';
+import ErrorMessage from './ErrorMessage';
 
 const propTypes = {
   className: PropTypes.string,
+  defaultValue: PropTypes.bool,
+  errorMessage: PropTypes.string,
   id: PropTypes.string.isRequired,
   innerRef: RefShape,
   isDisabled: PropTypes.bool,
@@ -15,16 +18,22 @@ const propTypes = {
   label: PropTypes.string.isRequired,
   labelClassName: PropTypes.string,
   name: PropTypes.string,
-  value: PropTypes.string,
+  onChange: PropTypes.func,
+  onSubmit: PropTypes.func,
+  value: PropTypes.bool,
   wrapperClassName: PropTypes.string,
 };
 const defaultProps = {
   className: null,
+  defaultValue: null,
+  errorMessage: null,
   isDisabled: false,
   isRequired: false,
   innerRef: null,
   labelClassName: null,
   name: null,
+  onChange: null,
+  onSubmit: null,
   value: null,
   wrapperClassName: null,
 };
@@ -32,6 +41,8 @@ const defaultProps = {
 /**
  * @param {Object} props
  * @param {string | null} [props.className]
+ * @param {boolean | null} [props.defaultValue]
+ * @param {string | null} [props.errorMessage]
  * @param {React.RefObject | null} [props.innerRef]
  * @param {string} props.id
  * @param {boolean} [props.isDisabled]
@@ -39,12 +50,16 @@ const defaultProps = {
  * @param {string} props.label
  * @param {string | null} [props.labelClassName]
  * @param {string | null} [props.name]
+ * @param {EventAction | null} [props.onChange]
+ * @param {(() => void) | null} [props.onSubmit]
  * @param {string | null} [props.value]
  * @param {string | null} [props.wrapperClassName]
  * @returns {JSX.Element}
  */
 export function RadioButton({
   className = null,
+  defaultValue,
+  errorMessage,
   id,
   isDisabled = false,
   isRequired = false,
@@ -52,14 +67,25 @@ export function RadioButton({
   label,
   labelClassName = null,
   name = null,
+  onChange,
+  onSubmit,
   value = null,
   wrapperClassName = null,
+  ...rest
 }) {
   const {
+    currentErrorMessage,
     currentOnChange,
-    currentValue,
     currentOnFormKeyPress,
-  } = useRadioButtonsContext();
+    currentValue,
+  } = useCurrentValuesFromForm({
+    defaultValue,
+    errorMessage,
+    id,
+    onChange,
+    onSubmit,
+    value,
+  });
 
   return (
     <div className={joinClassNames('input-wrapper input-wrapper--radio', wrapperClassName)}>
@@ -68,18 +94,20 @@ export function RadioButton({
         {isRequired ? <RequiredStar /> : null}
       </label>
       <input
-        checked={currentValue === value}
+        aria-describedby={currentErrorMessage ? `${id}-error` : null}
+        checked={currentValue}
         className={className}
         disabled={isDisabled}
         id={id}
-        name={name}
-        onChange={useCallback(currentOnChange, [currentOnChange])}
+        name={name || id}
+        onChange={currentOnChange}
         onKeyPress={currentOnFormKeyPress}
         ref={innerRef}
         required={isRequired}
         type="radio"
-        value={value}
+        {...rest}
       />
+      <ErrorMessage errorMessage={currentErrorMessage} id={id} />
     </div>
   );
 }
