--- @utahdts/utah-design-system/react/components/forms/radioButton/hooks/useRadioButtonsContext.jsx
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import { RadioButtonsContext } from '../util/RadioButtonsContext';
-
-export function useRadioButtonsContext() {
-  return useContext(RadioButtonsContext);
-}
