--- @utahdts/utah-design-system/react/components/forms/radioButton/util/RadioButtonsContext.jsx
@@ -1,9 +1,0 @@
-import { createContext } from 'react';
-
-/** @typedef {import('../../../jsDocTypes').RadioButtonsContextValue<any>} RadioButtonsContextValue */
-
-export const RadioButtonsContext = /** @type {typeof createContext<RadioButtonsContextValue>} */ (createContext)({
-  currentOnChange: null,
-  currentValue: null,
-  currentOnFormKeyPress: null,
-});
