+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/RadioButton/RadioButtonExampleRender.jsx
@@ -1,11 +1,12 @@
 // @ts-check
-import { RadioButton, RadioButtonsWrapper, RefShape } from '@utahdts/utah-design-system';
+import { RadioButton, RefShape } from '@utahdts/utah-design-system';
 import PropTypes from 'prop-types';
 import React from 'react';
 import RadioButtonExamplePropsShape from '../../../../../../propTypesShapes/RadioButtonExamplePropsShape';
 
 const propTypes = {
   innerRef: RefShape,
+  setState: PropTypes.func.isRequired,
   state: PropTypes.shape({
     props: RadioButtonExamplePropsShape.isRequired,
   }).isRequired,
@@ -15,6 +16,7 @@ const defaultProps = {
 };
 
 export function RadioButtonExampleRender({
+  setState,
   state: {
     props: {
       className,
@@ -24,29 +26,24 @@ export function RadioButtonExampleRender({
       isRequired,
       label,
       value,
-      wrapperClassName,
-      wrapperLabel,
     },
   },
   innerRef,
 }) {
   return (
     <div ref={innerRef}>
-      <RadioButtonsWrapper
-        className={wrapperClassName}
+      <RadioButton
+        className={className}
+        id={id || 'radio-button-example-render-id'}
+        label={label || 'Radio Label'}
         errorMessage={errorMessage}
-        label={wrapperLabel || 'Wrapper Label'}
-        id="radio-button-example-wrapper-id"
         isRequired={isRequired}
-      >
-        <RadioButton
-          className={className}
-          id={id || 'radio-button-example-render-id'}
-          isDisabled={isDisabled}
-          label={label || 'Radio Label'}
-          value={value || 'value example'}
-        />
-      </RadioButtonsWrapper>
+        isDisabled={isDisabled}
+        value={value}
+        onChange={() => setState((draftState) => {
+          draftState.props.value = !draftState.props.value;
+        })}
+      />
     </div>
   );
 }
