+++ utah-design-system-website/src/react/propTypesShapes/RadioButtonExamplePropsShape.js
@@ -9,7 +9,5 @@ export default PropTypes.exact({
   isDisabled: PropTypes.bool,
   isRequired: PropTypes.bool,
   label: PropTypes.string,
-  value: PropTypes.string,
-  wrapperClassName: PropTypes.string,
-  wrapperLabel: PropTypes.string,
+  value: PropTypes.bool,
 });
